fix potential problem with uncorrect `n_jobs` and move `determine_n_jobs` to `utilities`
